third_party/WebKit/LayoutTests/http/tests/inspector/network/network-worker-fetch-blocked.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    // Need to enable networkInWorkers experiment for off-main-thread-fetch virtual test.
    Runtime.experiments.enableForTest("networkInWorkers");

    InspectorTest.recordNetwork();
    SDK.multitargetNetworkManager.setBlockingEnabled(true);
    SDK.multitargetNetworkManager.setBlockedPatterns([{url: 'resources/resource.php', enabled: true}]);

    InspectorTest.makeFetchInWorker('resource.php', {}, fetchCallback);

    function fetchCallback(result)
    {
        InspectorTest.addResult('Fetch in worker result: ' + result.value);

        var requests = InspectorTest.networkRequests();
        requests.forEach((request) => {
          InspectorTest.addResult(request.url());
          InspectorTest.addResult("resource.type: " + request.resourceType());
          InspectorTest.addResult("request.failed: " + !!request.failed);
        });

        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests blocking fetch in worker.</p>
</body>
</html>
